+++ format_converter_test.go
@@ -18,7 +18,7 @@ func (s *CLISuite) TestRegexReplace(c *C) {
 	f := NewFormatConverter("(.*)/(.*)/DSC_(\\d+)\\.jpg$", "$1$2$3")
 
 	c.Check(f.String(test), Equals, "201609711")
-	c.Check(f.Convert(test), Equals, int64(201609711))
+	c.Check(f.Convert(test), Equals, uint64(201609711))
 }
 
 func (s *CLISuite) TestTimeReplace(c *C) {
@@ -27,5 +27,5 @@ func (s *CLISuite) TestTimeReplace(c *C) {
 	f.TimeLayout = "2006-01-02"
 
 	c.Check(f.String(test), Equals, "2016-09-22")
-	c.Check(f.Convert(test), Equals, time.Date(2016, time.September, 22, 0, 0, 0, 0, time.UTC).Unix())
+	c.Check(f.Convert(test), Equals, uint64(time.Date(2016, time.September, 22, 0, 0, 0, 0, time.UTC).Unix()))
 }
